requirements.txt
Django==5.0.6
djangorestframework==3.15.1
dj-database-url==2.1.0
openai==1.28.1
psycopg2-binary==2.9.9
